init monorepo packages
