+++ packages/applicant/handlers.ts
@@ -1,0 +1,3 @@
+export default class ApplicantProtocolHandlers {
+
+}
