+++ packages/applicant/index.ts
@@ -1,5 +1,5 @@
 export * from './config.js';
 export * from './handlers.js';
+export { protocol as credentialApplicantProtocol } from './protocol.js';
 export * from './server.js';
-
-export { protocol as credentialApplicantProtocol } from './credential-applicant.js';
+export * from './web5-manager.js';
