+++ packages/common/did-dht-manager.ts
@@ -8,7 +8,7 @@ import {
 } from '@web5/dids';
 import { FileSystem } from '../utils/file-system.js';
 import { Config } from './config.js';
-import { server } from './index.js';
+import IssuerServer from '../../issuer/server.js';
 
 /**
  * DidDhtManager handles interactions between the DCX server and the DID
@@ -51,7 +51,7 @@ export class DidDhtManager {
    * @returns DidRegistrationResult; see {@link DidRegistrationResult}
    */
   public async publishDidDoc(gatewayUri: string): Promise<DidRegistrationResult> {
-    gatewayUri ??= server.useOptions.gateways?.[0] ?? Config.DEFAULT_GATEWAY_URIS[0];
+    gatewayUri ??= IssuerServer.useOptions.gateways?.[0] ?? Config.DEFAULT_GATEWAY_URIS[0];
     return await DidDht.publish({ did: this.bearerDid, gatewayUri });
   }
