+++ packages/common/identity-vault.ts
