+++ packages/issuer/index.ts
@@ -1,5 +1,5 @@
-export { protocol as credentialIssuerProtocol } from './protocol/credential-issuer.js';
-
 export * from './config.js';
 export * from './handlers.js';
+export { protocol as credentialIssuerProtocol } from './protocol.js';
 export * from './server.js';
+export * from './web5-manager.js';
