--- src/protocol/index.ts
@@ -1,2 +1,0 @@
-export { protocol as credentialApplicantProtocol } from './credential-applicant.js';
-export { protocol as credentialIssuerProtocol } from './credential-issuer.js';
