--- src/schemas/index.ts
@@ -1,4 +1,0 @@
-export { schema as presentationSchema } from '../schemas/application.js';
-export { schema as responseSchema } from '../schemas/response.js';
-export { schema as invoiceSchema } from '../schemas/invoice.js';
-export { schema as manifestSchema } from '../schemas/manifest.js';
