--- src/schemas/invoice.ts
@@ -1,10 +1,0 @@
-export const schema = {
-  $schema    : 'http://json-schema.org/draft-07/schema#',
-  $id        : 'https://dcx.io/protocol/credential-issuance/schemas/invoice',
-  type       : 'object',
-  properties : {
-    url: {
-      type: 'string',
-    },
-  },
-};
