--- src/types/cipher.ts
@@ -1,4 +1,0 @@
-export enum CipherAlgorightm {
-  AES_256_CBC = 'aes-256-cbc',
-  AES_256_CTR = 'aes-256-ctr',
-}
