--- src/types/index.ts
@@ -1,4 +1,0 @@
-export type * from './cipher.js';
-export type * from './dcx.js';
-export type * from './did.js';
-export type * from './web5.js';
