--- src/utils/index.ts
@@ -1,8 +1,0 @@
-export * from './cipher.js';
-export * from './dwn.js';
-export * from './error.js';
-export * from './file-system.js';
-export * from './json.js';
-export * from './logger.js';
-export * from './objects.js';
-export * from './time.js';
