--- src/utils/string.ts
@@ -1,3 +1,0 @@
-export function isEmptyString(obj: unknown): obj is string {
-  return typeof obj !== 'string' || obj.trim().length === 0;
-}
